Plot interactive flow diagrams for systematic reviews

Produces a PRISMA2020 style flow diagram for systematic reviews, with the option to add interactivity through tooltips (mouseover popups) and hyperlink URLs to each box. Data can be imported from the standard CSV template provided.

PRISMA_flowdiagram(
  data,
  interactive = FALSE,
  previous = TRUE,
  other = TRUE,
  detail_databases = FALSE,
  detail_registers = FALSE,
  fontsize = 7,
  font = "Helvetica",
  title_colour = "Goldenrod1",
  greybox_colour = "Gainsboro",
  main_colour = "Black",
  arrow_colour = "Black",
  arrow_head = "normal",
  arrow_tail = "none",
  side_boxes = TRUE
)

Arguments

data

List of data inputs including numbers of studies, box text, tooltips, and urls for hyperlinks. Data inputted via the PRISMA_data() function. If inputting individually, see the necessary parameters listed in the PRISMA_data()) function and combine them in a list using data <- list().

interactive

Logical argument TRUE or FALSE whether to plot interactivity (tooltips and hyperlinked boxes).

previous

Logical argument (TRUE or FALSE) specifying whether previous studies were sought.

other

Logical argument (TRUE or FALSE) specifying whether other studies were sought.

detail_databases

Logical argument (TRUE or FALSE) specifying whether to list specific databases.

detail_registers

Logical argument (TRUE or FALSE) specifying whether to list specific registers.

fontsize

The font size for text in each box. The default is '12'.

font

The font for text in each box. The default is 'Helvetica'.

title_colour

The colour for the upper middle title box (new studies). The default is 'Goldenrod1'. See 'DiagrammeR' colour scheme. http://rich-iannone.github.io/DiagrammeR/graphviz_and_mermaid.html#colors.

greybox_colour

The colour for the left and right column boxes. The default is 'Gainsboro'. See 'DiagrammeR' colour scheme http://rich-iannone.github.io/DiagrammeR/graphviz_and_mermaid.html#colors.

main_colour

The colour for the main box borders. The default is 'Black'. See 'DiagrammeR' colour scheme http://rich-iannone.github.io/DiagrammeR/graphviz_and_mermaid.html#colors.

arrow_colour

The colour for the connecting lines. The default is 'Black'. See 'DiagrammeR' colour scheme http://rich-iannone.github.io/DiagrammeR/graphviz_and_mermaid.html#colors.

arrow_head

The head shape for the line connectors. The default is 'normal'. See DiagrammeR arrow shape specification http://rich-iannone.github.io/DiagrammeR/graphviz_and_mermaid.html#arrow-shapes. #nolint

arrow_tail

The tail shape for the line connectors. The default is 'none'. See DiagrammeR arrow shape specification http://rich-iannone.github.io/DiagrammeR/graphviz_and_mermaid.html#arrow-shapes. #nolint

side_boxes

Whether or not to include the blue label boxes along the side

Value

A flow diagram plot.

Examples

csvFile <- system.file("extdata", "PRISMA.csv", package = "PRISMA2020")
data <- read.csv(csvFile);
data <- PRISMA_data(data);
plot <- PRISMA_flowdiagram(data,
                fontsize = 12,
                interactive = TRUE,
                previous = FALSE,
                other = TRUE);
#> Warning: 1.5
plot