Save PRISMA2020 flow diagram

Save the output from PRISMA_flowdiagram() to the working directory.

PRISMA_save(
  plotobj,
  filename = "PRISMA2020_flowdiagram.html",
  filetype = NA,
  overwrite = FALSE
)

Arguments

plotobj

A plot produced using PRISMA_flowdiagram().

filename

The filename to save (including extension)

filetype

The filetype to save the plot in, supports: HTML, ZIP, PDF, PNG, SVG, PS and WEBP (if NA, the filetype will be calculated out based on the file extension) HTML files maintain hyperlinks and tooltips The ZIP option creates an archive containing the HTML file, alongside supporting javascript and css files in an adjacent folder, instead of embedded base64 within the HTML file

overwrite

if TRUE, will overwrite an existing file

Value

the absolute filename of the saved diagram plot.

Examples

csvFile <- system.file("extdata", "PRISMA.csv", package = "PRISMA2020")
data <- read.csv(csvFile);
data <- PRISMA_data(data);
plot <- PRISMA_flowdiagram(data,
                fontsize = 12,
                interactive = TRUE,
                previous = FALSE,
                other = TRUE);
#> Warning: 1.5
PRISMA_save(plot, filename = tempfile(), filetype="html");
#> [1] "/tmp/RtmpRPmMKF/file174b1708c871"